requirements.txt
pyTelegramBotAPI
python-dotenv
bs4
plotly.express